Improve yield+scan exec performance test
Before this commit the filters running in the background had a very low
chance of returning data.  This did not execercise the use case of an
iterator returning data very slowly.  Returning data very slowly results
in slowly filling a buffer.  The test was modified to run the filter
scans with multiple probabilites which causes data to return with
different velocities.  Also the yield filter was modfied to support
yielding across returning a key/value.

The design of the yield filter was improved so that subclasses do not
have to override init().