Terraform configurations for creating test infrastructure (#185)
This change includes Terraform configurations for creating and destroying compute resources for testing on AWS and Azure. The configurations install ZooKeeper, Hadoop, Accumulo, and Accumulo-Testing. Users can supply options for the versions of the software that should be installed or can supply their own binary tarballs for installation. See the README for detailed documentation.


Co-authored-by: Brian Loss <brianloss@gmail.com>
Co-authored-by: domgarguilo <dominic.garguilo@gmail.com>