fixes bug with bulk RW file partition point creation (#236)
There was code that did the following

    TreeSet<Integer> startRows = new TreeSet<>();
    startRows.add(0);
    while (startRows.size() < parts)
      startRows.add(rand.nextInt(LOTS));

The above code was replaced in 7453c37 with a stream. The stream
did not fully capture the original behavior of the loop.  This
change makes the stream fully capture that behavior.  Need to
ensure that `parts` unique random numbers are generated including
zero (like if the random number generator returns zero it should
be properly deduplicated).  The stream was not properly handling
the RNG returning duplicates or zero.