Updates to Terraform configuration based on recent changes (#270)
This commit bumps the OpenTelemetry and Micrometer versions to
match the versions in the current 2.1 branch pom.xml file. This
also adds variables for os distribution and version to that special
handling could be added to install both Java 11 and Java 17 (recent
changes to Accumulo require Java 17 to build). Accumulo still uses
Java 11 at runtime. Changes tested with CentOS7 and RHEL8.