fixes bug in bulk batch writer
BulkBatchWriter was using Arrays.compare to compare Accumulo rows which
is incorrect because it compares bytes as signed integers.  Need to
compare the bytes as unsigned integers.  Updated the code to correctly
compare.  This bug would not cause data to go to the wrong tablet it
could only cause needless creation of many files for a single tablet.