fixes sorting in bulk batch writer (#289)
BulkBatchWriter was sorting only on the mutation row.  Occasionally
continuous ingest would generate the same 64 bit id twice in the same
batch and if its columns were not sorted properly then the write to the
rfile would fail.

This fixes the problem by sorting on the entire key instead of only the
row.